+++ src/conditional-visibility.ts
@@ -141,15 +141,12 @@ Hooks.once('libChangelogsReady', function () {
   libChangelogs.register(
     CONSTANTS.MODULE_NAME,
     `
-    - Integration autostealth with better rolls
-    - Resolved winking by disable the 'SightLayer.prototype.tokenVision' wrapper (maybe ???)
-    - Update the checker again
-    - Give up on understand how the community want to manage the passive stealth on Dnd5e, because of this i create a new active effect 'Stealthed' so everyone can apply is own custom solution. So now you can apply any combination of 'Hidden' vs 'Stealthed' and simply disable or remove the active effect if you don't want to check the visibility. For moredetails check out the token configuration panel.
-    - Update compatibility with new 'Levels' losHeight function
-    - Buf fix for 'Auto Stealth' and update token vs update actor
-    - Add a clean up macro on the module for remove the old flag per midi integration
-    - [BREAKING CHANGES] You must reset the module settings with the dedicate button on the module settings panel for apply the new active effect stealthed
-    - Added  a new experimental module setting "Fast Mode" module setting, for anyone have some latency problem ? With this the only the time used for the check is the time used for find a value on a map javascript with a key (so it should be faster).
+    - Add 'ATCV.conditionBlinded' and 'ATCV.conditionBlindedOverride' condtions for [https://github.com/p4535992/conditional-visibility/issues/32](https://github.com/p4535992/conditional-visibility/issues/32)
+    - Clean up comments
+    - Add module settins for hud position [Option for changing CV's HUD button position](https://github.com/p4535992/conditional-visibility/issues/34)
+    - Bug fix: [[BUG] Duplicate token configuration panel and some sight issue bug](https://github.com/p4535992/conditional-visibility/issues/29)
+    - Add class field to extra_sense template
+    - Update renderChat for autostealth for mutliligual (the community need to pr me a language file)
     `,
     'minor',
   );